% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pmprs_x}
\alias{pmprs_x}
\title{Title}
\usage{
pmprs_x(b_exp, maf_x)
}
\arguments{
\item{b_exp}{a vector of SNP effects on the exposure}

\item{maf_x}{minor allele frequency}
}
\value{
population mean polygenic risk scores
}
\description{
Title
}
