% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pmprs_y}
\alias{pmprs_y}
\title{Title}
\usage{
pmprs_y(b_out, maf_y)
}
\arguments{
\item{b_out}{a vector of SNP effects on the outcome}

\item{maf_y}{minor allele frequency}
}
\value{
population mean polygenic risk scores
}
\description{
Title
}
