% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{RSR}
\alias{RSR}
\title{Title}
\usage{
RSR(b_exp, b_out, se_exp, se_out, maf_x, maf_y)
}
\arguments{
\item{b_exp}{a vector of SNP effects on exposure}

\item{b_out}{a vector of SNP effects on the outcome}

\item{se_exp}{a vector of stander error of b_exp}

\item{se_out}{a vector of stander error of b_out}

\item{maf_x}{minor allele frequency}

\item{maf_y}{minor allele frequency}
}
\value{
risk score ratio
}
\description{
Title
}
