% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{var_x}
\alias{var_x}
\title{Title}
\usage{
var_x(b_exp, maf_x, se_exp)
}
\arguments{
\item{b_exp}{a vector of SNP effects on the exposure}

\item{maf_x}{minor allele frequency}

\item{se_exp}{a vector of stander error of b_exp}
}
\value{
the variance of pmprs_x
}
\description{
Title
}
