% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{var_y}
\alias{var_y}
\title{Title}
\usage{
var_y(b_out, maf_y, se_out)
}
\arguments{
\item{b_out}{a vector of SNP effects on the outcome}

\item{maf_y}{minor allele frequency}

\item{se_out}{a vector of stander error of b_out}
}
\value{
the variance of pmprs_y
}
\description{
Title
}
